Update rating variable to be double instead of int
Everything works fine if a user adds a rating that is an integer. If a user adds a new app rating that is a double, there is one extra star. This is due to updateButtonSelectionStates() in RatingControl.swift.

+++ ExpenseTracker/AddAppViewController.swift
@@ -102,7 +102,7 @@ class AddAppViewController: UIViewController, UITextFieldDelegate, UIImagePicker
         
         let name = nameTextField.text ?? ""
         let photo = photoImageView.image
-        let rating = Int(ratingTextField.text!)!
+        let rating = Double(ratingTextField.text!)!
         let date = dateTextField.text ?? ""
         
         // Set the app to be passed to AppTableViewController after the unwind segue
